getMessages: support initialRequest, move message refresh to fimStandard from fimApi
In fimApi, initialRequest is a flag that sets the appropriate flags for
an initial request
In fimStandard, requestMessages now uses getMessages.php for the first
request, and stream.php for later requests.
requestMessages, when not using stream.php, will handle the refresh
queries itself, running them every 5 seconds, and slowing down on
non-exception failures. It will not refresh if
requestSettings.serverSentEvents is true, which may happen after several
requests.