Database: add decode data transformation support (and also allow unlimited array sizes for $db->getAsArray([])
I am concerned about the performance of this, and somehow still haven't
created unit tests that could be helpful. Later on, I'll want to check
it fully, and possibly make optimisations if desired.

Meanwhile, as part of this change the array building logic in
getAsArray() was rewritten to use references, and it should now support
unlimited array sizes, whereas before it was capped at a depth of 5. (A
benefit of learning C pointers, as it happens.)