WebPro: Rewrite to eliminate XSS vulnerabilities when displaying messages.
This is the first of several rewrites that will have to occur to remove XSS vulnerabilities. At present, FreezeMessenger's WebPro frontend is entirely and completely suspectiple to XSS from altering the delivered JSON. It is likely even a little suspecitible to, e.g. a username with HTML formatting in it. Both are being fixed, but along with other changes.