Remove unused functions from fim_general.php.
Most were left over from when we were doing templating server-side (...which was a crazy long time ago). 

Additionally, we rewrite the cache function for getting configuration data to use json_decode instead of our weird decode function when dealing with arrays. I... don't know where the corresponding encode is even performed, but it will be broken for now. (Honestly, probably already was).