Database: Add hard partitioning
Hard partitioning makes multiple copies of the same table, and then attempts to automatically select the correct table based on the partition column, as defined at runtime.

It's a bit complicated, and won't work universally; I've done my best to include error messages where hard partioning isn't supported.

It is currently used on messagesCache and messages, both with a partition count of 10, and on roomId. It seems to work, amazingly, but some edge cases probaly remain.