Schema: make memory table message length 1000 (the default maximum)
Having it 5000 would, if the defaults are kept, waste 4KB of memory (!) per message stored. Really, really dumb.

(Incidentally, the memory table is at least part of the reason for the low default. There's really no substantive reason for it to be so low, except that the memory table is going to allocate the full Kilobyte whether it needs it or not. On PostGres, a higher default should be attempted.)