Collection Triggers! Glorious collection triggers!
These are somewhat silly, but serve a very useful purpose: maintaining list caches of aggregate tables, e.g. a user's watchRooms list.

There's a lot of potential here. While we currently use this to keep the in-data table summary caches up-to-date, it will eventually be used for things like Redis as well. It's a great feature for the IM, and shouldn't result in too much further instability, since the stuff that will be using it was already broken.

Additionally, at least watch rooms is now correctly stored and retrieved.