Rudimentary caching of fimUser objects
Caches are set on __destruct.

Average cache size (with this change) is about 2kb. Quite a few users should be storable here, though it's hardly required (even without any caching, the basic user operations are rather well optimised.)

Currently caches for 5 minutes (settable in defaultConfig); could probably be very high, but 5 minutes will cover most situations where a fimUser object is called (that is, between active session API calls).

Does NOT cache list changes yet. As such, editUserSettings is currently sorta broken -- when a change is made, it will not be reflected in getUsers until the cache expires. Which, for a logged-in user, could be a while.