Add basic JavaGUI prototype.
What works? Well, login, kinda, though not through a GUI, and it currently uses hard-coded values.

But there's a reasonably good foundation here for HTTP requests, and JSON parsing seems to be in order. Both of these are meaningful.

Maven dependencies will be required for the above to work. In IDEA, open pom.xml and it will usually show a popup in the bottom right of the screen asking if you want to import the deprencies.