api/getMessages: no longer return user data
This is a logical choice (as it reduces server load and makes certain design a fair bit easier).

The same convention is followed by stream.php, and is respected by both WebPro and JavaGUI. WebPro updated its deferred user/room data code to ensure that it doesn't repeat requests for the same information, which would now otherwise be happening.

The database itself is NOT yet updated. The join in getMessages() and the denormalisation in the messagesCached table must both be removed.