API: merge editUserStatus and getActiveUsers into userStatus
WebPro updated to match. Documentation updated. "onlineThreshold" removed from getActiveUsers -- there is no real reason for clients to use it, and it could make caching harder.

Also:
'POST' requests are now treated as 'create' requests, and 'PUT' requests are now treated as 'edit' requests (used to be the opposite). WebPro updated to match. This is more in-line with the REST standard (where PUT is idempotent). WebPro is also no longer going to use any PUT and DELETE requests, since not all web servers support them.