unreadMessages unit tests; move list changes to fimUser; add list cache trigger for fimRoom
Not quite fully working yet, but close. The watchedByUsers cache can't be effectively updated (in part because it's too easy for a single user to cause too many locks on the rooms table), and will be put on a timed cache instead.

I'm also not entirely certain about keeping the list triggers in fimUser, but doing so likely makes fimUser object caching more reliable. Not to mention, it's just a better assignment of responsibilities in general.