Backend: (Mostly) Working PHPBB Logins
Note, somewhat to my chagrin, that PHPBB no longer appears to have a completely standard hash format. As installed on my system, it is using hashes that are compatible with PHPAss, and hopefully it will continue to do so. Still, not all existing PHPBB installations will necessarily work.

Anyway, the notable changes:
1.) Fix $integrationDatabase declaration in global.php
2.) Pass DatabaseSQL instance to LoginFactory
3.) Clean up cleanSessions() function
4.) Small cleanup to IntegrationLogin GrantType
5.) Add LoginPhpbb Class
6.) Remove PHPBB hash code from fimUser
7.) Start removing code from fimDatabaseUAC; stop using the class entirely (will delete once its code is no longer useful as reference)
8.) Temporarily disable emoticons entirely. Will need to decide how to enable those on PHPBB/vBulletin (probably through method in LoginFactory to expose active primary LoginRunner, which will then have getSmilies method exposed through as abstract in LoginDatabase.)

In any case, the login system is starting to look quite nice and robust. I'm very happy with the reflection techniques it is now using; it should be very easy to slot in new priamry and secondary LoginRunners, without requiring code changes (except to config.php).