Stress Test "Proof of Concept"
This also fixes some small bugs encountered while creating the PoC, and begins a process I've intended for a while: ensuring all fimApi calls return deferreds.

In any case, the proof of concepts right now does the following:
1.) Creates a bunch of users.
2.) Creates a couple of rooms for them.
3.) Has those users send and receive messages in the rooms.

It will be expanded to use stream.php for half of the users, as well as pinging/active users queries (and eventually other stuff, but that's the baseline goal for now).