Some legacy code removal and other minor cleanup.
Amazingly, PHP 5.2 was still the stated requirement. For development, that will now be PHP 7; I will be looking into transpiling back to PHP 5.5/5.6 for release. Before that is EOL, so I've no interest in supporting.

Note, this commits removes custom JSON and Sha256 functions. Both are now available in basicaly every supported PHP release, and the custom handlers add cruft and are slower.