Backend: Rewrite generalCache to use CacheInterface; namespace
Right now, this should be drop-in compatible with the old generalCache file, but that will change once cleanup occurs.

Well, with one exception: the local-memory store has been removed. If no cache method is available, using set followed by get will not return the set value. So that is no longer guaranteed to be valid.