Rudimentary population of user groups with LoginReddit, LoginSteam
Pretty much the main reason I added them (and something I thought of literally as I was sleeping, which is actually pretty rare for me, personally), it's looking fairly promising.

Usergroup functionality won't be that hard, I don't believe. We're really only exposing it in three places (hasHermission, setPermission, and getGroups), and most of the functionality already exists in some form.

This also updates the schema to support those inserts, and makes some minor fixes. Finally, it disables database error handling when error_reporting() == 0, which... well, I'll need to flesh that out, probably. I'm honestly not sure off hand if it's normally 0 or not.