DBAL: SqlServer support
This is not yet comphrensive (upsert on columns with serial are not working yet), but installation and sending messages works as expected.

It results in some kinda major changes to our underlying database assumptions, as well:
1.) Serial columns should always start at 1 from now on. A DBMS that doesn't do that needs to be worked around.
2.) Manually specifying serial columns at insert should be avoided. We do it once to add the root users, but will hopefully not do so otherwise. (It requires a workaround for SqlServer when we do).

Note that this is NOT INJECTION SAFE yet. SqlServer does not have its own escape function, and our implementation is not yet exhaustively injection proof.

Some smaller tweaks were made to the MySQL implementation for compatibility.