dbSchema.xml: A first pass at optimising indexes.
These changes are all theoretical, but no real thought has gone into index design since the original schema was created. So it's a bit overdue after all the other changes.

This also adds markers for overriding to btree/hash index storage, which is not yet implemented.

It also considers the order of indexes for btree storage (would you believe I didn't know it mattered until... a week ago?)

(It is an untested change, so it probably breaks installation in at least one way.)