WebPro/Room: Detect event stream failure
I was unclear on the rules before; after a bit of investigation, you can tell whether or not an error will cause the event sourcce to stop retrying, and it's only in these instances you want to switch to fallback mode anyway (e.g. when a 500 error occurs).

So yay, this is all pretty solid now.