WebPro: "bubble" formatting option
Frankly, it just doesn't look good without the bubble. So I'll likely either need to tweak the non-bubble formatting further, or remove the setting.

(The main benefit of the setting is on slower devices, where render the bubbles may be difficult. I'm not positive this is actually a concern, however.)