WebPro: Unread message count
Pretty self-explanatory: the number of unread messages will now be counted. This should both make it clearer that a bolded entry in the room list means new messages, and allow users to quickly determine how many messages there are.

Feature-wise, the only remaining improvement here is to jump to the last unread message when a room is opened.