DatabaseInstance\createUnreadMessage: Only for logged-in users
While this particular issue hasn't yet come up in testing, it would no doubt become a _major_ chokepoint when rooms with thousands of following users suddenly need to create watch events.

This does, in-fact, remain one of a handful of chokepoints where a bunch of rows are upserted in sequence (the other major one being at login). In truth, I don't see a great solution for this; the best option may be to remove the "otherMessages" column and simply do a batch of inserts, assuming the database is kind enough to not stop on the first one that fails. The better option, and one I will likely persue in Beta 4, would be:
1.) Create/delete unreadMessage table entires whenever a user subscribes/unsubscribes
2.) Do a bulk update on this table, e.g. UPDATE unreadMessages SET totalMessages = totalMessages + 1, lastMessage = X WHERE userId IN (1, 2, 3, ...)
3.) To select unread messages, select rows with totalMessage > 0. (With proper indexes, this should be easy enough.)

While this solution would have a more linear scaling, it would still most likely be inefficient for small-to-medium installations, and it needs more consideration.