WebPro\Service Workers: Progressive loading
Service workers will now only be enabled when push messaging is enabled (though in Beta 4 they will hopefully be enabled by deafult in supported browsers)

They will also only _register_ -- that is, use the service worker functionality -- if a controller is available immediately, and no longer attempt to take control of the page at initial installation. If a controller isn't available, the fallback mode will be used.

Finally, both the push notification option and the regulat notification option will be hidden from users if the functionality is not supported.